+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added: Vertex.query limited by connected edge labels
 * Fixed: Memory leak when using compression in serializer
+* Speed up Accumulo EdgeInfo and fix null conditions
 
 # v2.4.4